backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Telegram Bot
aiogram==3.4.1
pydantic-settings==2.1.0

# Async HTTP
aiohttp==3.9.1
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
